Add Stat. Fix grid
