+++ .idea/misc.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="JavaScriptSettings">
-    <option name="languageLevel" value="JSX" />
-  </component>
+    <component name="JavaScriptSettings">
+        <option name="languageLevel" value="FLOW"/>
+    </component>
 </project>
