+++ package.json
@@ -10,6 +10,7 @@
     "date-fns": "^2.16.1",
     "es7": "^1.0.8-beta.4",
     "history": "^5.0.0",
+    "moment": "^2.29.1",
     "react": "^17.0.1",
     "react-dates": "^21.8.0",
     "react-dom": "^17.0.1",
