--- src/pages/Calendar.css
@@ -1,115 +1,0 @@
-@import url('http://fonts.googleapis.com/css?family=Lato&subset=latin,latin-ext');
-
-* {
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    font-family: 'Lato', sans-serif;
-}
-
-.booking_schedule {
-    display: grid;
-    grid-column-start: 2;
-    grid-column-end: 16;
-    grid-row-start: 1;
-    height: 100vh;
-    align-items: center;
-    justify-content: center;
-    font-size: 3rem;
-    background-color: #b2c2d7;
-}
-
-
-.arrival_schedule{
-    padding: 60px;
-    display: grid;
-    grid-column-start: 2;
-    grid-column-end: 16;
-    grid-row-start: 1;
-    height: 50vh;
-    align-items: center;
-    justify-content: center;
-    font-size: 1rem;
-
-}
-
-
-.spoil{
-
-  display: flex;
-    justify-content: space-around;
-    align-items: flex-start;
-padding-right: 10px;
-padding-left: 10px;
-flex-direction: row;
-z-index: 0;
-}
-
-.spoil-inside{
-
-  display: flex;
-  align-content: space-between;
-
-
-
-
-flex-wrap: wrap;
-
-
-}
-.detblock{
-  display: flex;
- align-items: center;
- justify-content: center;
- margin-bottom: 25px;
- margin-top: 15px;
-}
-.spoilb{
-
-      background:#f7f7f7;
-z-index: 0;
-padding-top: 30px;
-
-}
-.calendarr{
-
-      background:#ededed;
-
-
-}
-.links{
-    font-size: 1rem;
-      color: #00a699;
-
-}
-
-.link{
-
-  border-radius:5px;
-  background:#00a699;
-      color:#fff;
-      box-shadow: 0 3px 0 #0c8d72;
-  border: 2px solid transparent;
-margin-left: 8px;
-
-  font-size: 1rem;
-  font-weight: 600;
-  line-height: 1;
-  margin-top: 15px;
-  width: 104px;
-  text-align: center;
-  text-shadow:0 1px 1px rgba(0, 0, 0, 0.5);
-}
-
-.page{
-    display: grid;
-    grid-template-columns: repeat(15, 1fr);
-    grid-auto-rows: minmax(100px, auto);
-}
-.sidebar{
-    position: relative;
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 1;
-    grid-row-end:3;
-}
