--- src/pages/Statistics.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-
-class Statistics extends Component {
-    render() {
-        return (
-            <div className={'statistics'}>
-                <h1>Statistics</h1>
-            </div>
-        );
-    }
-}
-
-export default Statistics;
