+++ src/pages/accommodations/Accommodations.css
@@ -1,5 +1,9 @@
 .accommodations {
-background-color: whitesmoke;
+    grid-column-start: 2;
+    grid-column-end: 16;
+    height: 100vh;
+    margin-left: auto;
+    margin-right: auto;
 }
 
 .rectangle_accommodations {
@@ -12,14 +16,14 @@ background-color: whitesmoke;
     border-radius: 20px;
     border-style: solid;
     border-color: rgba(224, 228, 231, 0.836);
-    
+
     transition: all linear .2s;
 }
 
 .rectangle_accommodations:hover {
     border-color: white;
     box-shadow: 0 14px 28px rgba(3, 69, 131, 0.603), 0 4px 12px rgba(36, 131, 219, 0.842);
-   
+
 }
 
 .vertical {
@@ -31,31 +35,31 @@ background-color: whitesmoke;
     display: flex;
     flex-direction: row;
 }
+
 .textInCards_accommodations {
-    
-    font-family: Roboto;
+    font-family: Roboto, serif;
     font-style: normal;
     font-weight: 400;
     font-size: 12px;
     line-height: 12px;
-  
+
     text-align: center;
     letter-spacing: 1.25px;
     text-transform: uppercase;
     color: #000000;
 }
 
-.image_accommodations{ 
+.image_accommodations {
     margin-top: 10px;
     margin-left: 10px;
     margin-right: 10px;
     width: 314px;
     height: 200px;
     border-radius: 20px;
-   
+
 }
 
-.buttonAddNewHotel_accommodations{
+.buttonAddNewHotel_accommodations {
     width: 304px;
     height: 342px;
     margin: 20px;
@@ -73,17 +77,18 @@ background-color: whitesmoke;
 
     transition: all linear .2s;
 }
+
 .buttonAddNewHotel_accommodations:hover {
     font-size: 18px;
-    
+
     border-color: white;
 
     background: rgba(40, 125, 204, 0.973);
     box-shadow: 0 4px 15px rgba(36, 131, 219, 0.842);
-   
+
 }
 
-.hotelSettingsButtom_accommodations{
+.hotelSettingsButtom_accommodations {
     margin-left: 30px;
     margin-bottom: 40px;
     font-size: 12px;
@@ -99,7 +104,7 @@ background-color: whitesmoke;
 }
 
 
-.roomsButton_accommodations{
+.roomsButton_accommodations {
     margin-left: 20px;
     margin-bottom: 40px;
     font-size: 12px;
@@ -107,21 +112,21 @@ background-color: whitesmoke;
     letter-spacing: 1.25px;
     text-transform: uppercase;
     color: #000000;
-    
+
     background-color: white;
     border-color: white;
 
     transition: all linear .2s;
 }
 
-.hotelSettingsButtom_accommodations:hover , .roomsButton_accommodations:hover {
+.hotelSettingsButtom_accommodations:hover, .roomsButton_accommodations:hover {
     font-size: 14px;
-     background: rgba(40, 125, 204, 0.973);
-   
+    background: rgba(40, 125, 204, 0.973);
+
 }
 
 
-.button_link{
-   
+.button_link {
+
     margin-top: 0px;
 }
