+++ src/pages/accommodations/AddNewHotel.css
@@ -1,8 +1,12 @@
-.addNewHotel {
-    background-color: whitesmoke;
-    width: 100%;
-    height: 100%;
+.add_new_hotel {
+    grid-column-start: 2;
+    grid-column-end: 16;
+    grid-row-start: 1;
+    height: 100vh;
+    margin-left: auto;
+    margin-right: auto;
 }
+
 .title_new_hotel {
     margin-top: 45px;
     margin-left: 250px;
@@ -15,21 +19,21 @@
     letter-spacing: 0.15px;
     color: #000000;
 }
+
 .rectangle_new_hotel {
     width: 1000px;
     height: 600px;
-    margin-left: 350px;
-    margin-top: 100px;
     background: white;
     border-radius: 20px;
     border-style: solid;
     border-color: rgba(224, 228, 231, 0.836);
     transition: all linear .2s;
     box-shadow: 0 14px 28px rgba(3, 69, 131, 0.603), 0 4px 12px rgba(36, 131, 219, 0.842);
+
 }
 
 .card_new_hotel {
-    margin-top: 100px;
+    margin-top: 5rem;
 }
 
 
@@ -57,19 +61,19 @@ div h1.enterHotelData_new_hotel {
 input.input_data_name {
     width: 500px;
     height: 50px;
-   
+
 }
 
 input.input_data_adress {
     width: 500px;
     height: 50px;
-  
+
 }
 
 input.input_data_rooms {
     width: 200px;
     height: 50px;
-   
+
 }
 
 .button_save_new_hotel {
