+++ src/pages/home_page.css
@@ -7,9 +7,8 @@
     font-family: 'Lato', sans-serif;
 }
 
-.accommodations,
 .arrival_schedule, .booking_schedule,
-.settings, .statistics, .logout {
+.settings {
     display: grid;
     grid-column-start: 2;
     grid-column-end: 16;
