+++ src/pages/statistics/statistics.css
@@ -1,0 +1,28 @@
+.statistics {
+    display: grid;
+    grid-column-start: 2;
+    grid-column-end: 16;
+    height: 100vh;
+    margin: 1rem;
+}
+
+.stat_title {
+    grid-row: 1;
+    grid-column: 2;
+    justify-self: center;
+}
+
+.stat_proceeds {
+    position: relative;
+    grid-row: 1;
+    grid-column: 2;
+    margin: 1rem;
+
+}
+
+.mini_graphs {
+    grid-row: 1;
+    grid-column: 3;
+    margin: 1rem;
+}
+
