--- src/pages/tooltip.css
@@ -1,111 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Raleway:200,200i,400,400i,500,500i,700,700i');
-/*tipped element. should be inline-block or block*/
-.qtip {
-  display: inline-block;
-  position: relative;
-  cursor: pointer;
-  color: #fff;
-  border-bottom: 0.05em dotted #3bb4e5;
-  box-sizing: border-box;
-  font-style: normal;
-  transition:all .25s ease-in-out
-}
-.qtip:hover {color:#fff;border-bottom:0.05em dotted #069}
-/*the tip*/
-.qtip:before {
-  content: attr(data-tip);
-  font-size: 10px;
-  position: absolute;
-  background: rgba(10, 20, 30, 0.85);
-  color: #fff;
-  line-height: 1.2em;
-  padding: 0.5em;
-  font-style: normal;
-    z-index: 10000;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.3);
-  width: 114px;
-  text-align: center;
-  opacity: 0;
-  visibility: hidden;
-  transition: all .3s ease-in-out;
-  text-shadow: 1px 1px 1px rgba(0, 0, 0, 0.5);
-  font-family: sans-serif;
-  letter-spacing: 0;
-  font-weight: 50
-}
-.qtip:after {
-  width: 0;
-  height: 0;
-  border-style: solid;
-  content: '';
-  position: absolute;
-  opacity: 0;
-  visibility: hidden;
-  transition: all .3s ease-in-out
-}
-.qtip:hover:before,
-.qtip:hover:after {
-  visibility: visible;
-  opacity: 1
-}
-/*top*/
-.qtip.tip-top:before {
-  top: 0;
-  left: 50%;
-  transform: translate(-50%, calc(-100% - 8px));
-  box-sizing: border-box;
-  border-radius: 3px;
-}
-.qtip.tip-top:after {
-  border-width: 8px 8px 0 8px;
-  border-color: rgba(10, 20, 30, 0.85) transparent transparent transparent;
-  top: -8px;
-  left: 50%;
-  transform: translate(-50%, 0);
-}
-/*bottom*/
-.qtip.tip-bottom:before {
-  bottom: 0;
-  left: 50%;
-  transform: translate(-50%, calc(100% + 8px));
-  box-sizing: border-box;
-  border-radius: 3px;
-}
-.qtip.tip-bottom:after {
-  border-width: 0 8px 8px 8px;
-  border-color: transparent transparent rgba(10, 20, 30, 0.85) transparent;
-  bottom: -8px;
-  left: 50%;
-  transform: translate(-50%, 0);
-}
-/*left*/
-.qtip.tip-left:before {
-  left: 0;
-  top: 50%;
-  transform: translate(calc(-100% - 8px), -50%);
-  box-sizing: border-box;
-  border-radius: 3px;
-}
-.qtip.tip-left:after {
-  border-width: 8px 0 8px 8px;
-  border-color: transparent transparent transparent rgba(10, 20, 30, 0.85);
-  left: -8px;
-  top: 50%;
-  transform: translate(0, -50%);
-}
-/*right*/
-.qtip.tip-right:before {
-  right: 0;
-  top: 50%;
-  transform: translate(calc(100% + 8px), -50%);
-  box-sizing: border-box;
-  border-radius: 3px;
-}
-.qtip.tip-right:after {
-  border-width: 8px 8px 8px 0;
-  border-color: transparent rgba(10, 20, 30, 0.85) transparent transparent;
-  right: -8px;
-  top: 50%;
-  transform: translate(0, -50%);
-}
-/*some styles for this example*/
